requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
anyio
APScheduler
arrow
art
attrs
beautifulsoup4
binaryornot
blinker
Brotli
bs4
cachetools
certifi
cffi
chardet
charset-normalizer
click
cloudscraper
colorama
cookiecutter
cookiejar
cryptography
decorator
Deprecated
dnspython
docutils
fake-useragent
Faker
filetype
Flask
flask-cors
Flask-Limiter
frozenlist
ftfy
future
h11
h2
hpack
html5lib
httpcore
httpx
huepy
hypercli
hyperframe
idna
imageio
imageio-ffmpeg
iniconfig
instabot
instaloader
itsdangerous
Jinja2
kaitaistruct
Kivy
kivy-deps.angle
kivy-deps.glew
Kivy-Garden
kivy_deps.sdl2
kivymd
limits
MailTm
markdown-it-py
MarkupSafe
mdurl
mechanize
mock
MouseInfo
moviepy
multidict
numpy
opencv-python
ordered-set
outcome
packaging
pager
pillow
pluggy
proglog
propcache
psutil
pure-python-adb
pyasn1
PyAutoGUI
pycparser
pycryptodome
pydivert
pyfiglet
PyGetWindow
Pygments
PyMsgBox
pynput
pyOpenSSL
pyparsing
pyperclip
pypiwin32
PyQt6
PyQt6-Qt6
PyQt6_sip
PyRect
PyScreeze
PySocks
pystyle
pyTelegramBotAPI
pytest
python-dateutil
python-dotenv
python-slugify
python-telegram-bot
pytweening
pytz
pywin32
PyYAML
qrcode
random-user-agent
requests
requests-toolbelt
responses
rich
schedule
selenium
selenium-wire
setuptools
six
sniffio
sortedcontainers
soupsieve
tabulate
telebot
telegram
termcolor
text-unidecode
tornado
tqdm
trio
trio-websocket
types-python-dateutil
typing_extensions
tzdata
tzlocal
undetected-chromedriver
urllib3
wcwidth
webdriver-manager
webencodings
websocket-client
websockets
Werkzeug
wrapt
wsproto
yarl
zstandard
